Refactor of tests fir GHA
