+++ instrumentation/aws-java-sdk-dax-2.0.0/src/test/java/com/nr/instrumentation/dynamodb_dax/DAXClientConfigCacheTest.java
@@ -5,7 +5,7 @@
  *
  */
 
-package com.nr.instrumentation.dynamodb_v2;
+package com.nr.instrumentation.dynamodb_dax;
 
 import org.junit.Test;
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
