+++ instrumentation/aws-java-sdk-dax-2.0.0/src/test/resources/mockito-extensions/org.mockito.plugins.MockMaker
@@ -1,0 +1,1 @@
+mock-maker-inline
